Nameserver resolving overhaul
